rename ranker_obj file to ranking
